+++ semiproject/src/com/Writing/CustomInfo.java
@@ -6,9 +6,9 @@ public class CustomInfo {
 
 	private String userName;
 
-	//ï¿½Î±ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ Session ï¿½ï¿½ id, name ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ ï¿½Ã·ï¿½ï¿½Î°ï¿½, index.jspï¿½ï¿½ redirect ï¿½ï¿½ï¿½ï¿½. 
+	//?¥á??? ?????? Session ?? id, name ?????? ?¡À????, index.jsp?? redirect ????. 
 
-	//ï¿½ï¿½ï¿½Ç¿ï¿½ ï¿½Ã¸ï¿½ ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ idï¿½ï¿½ nameï¿½ï¿½ ï¿½ï¿½ï¿½ Å¬ï¿½ï¿½ï¿½ï¿½(CustomInfo) ï¿½ï¿½ï¿½ï¿½
+	//????? ?©ª? ?????? id?? name?? ??? ?????(CustomInfo) ????
 
 	public String getUserId() {
 		return userId;
