+++ semiproject/src/com/pDetail/PdetailDTO.java
@@ -1,0 +1,51 @@
+package com.pDetail;
+
+public class PdetailDTO {
+
+	private int num;
+	private String name;
+	private int price;
+	private int pro_size;
+	private int color;
+	private int count;
+	
+	public int getNum() {
+		return num;
+	}
+	public void setNum(int num) {
+		this.num = num;
+	}
+	public String getName() {
+		return name;
+	}
+	public void setName(String name) {
+		this.name = name;
+	}
+	public int getPrice() {
+		return price;
+	}
+	public void setPrice(int price) {
+		this.price = price;
+	}
+	public int getSize() {
+		return pro_size;
+	}
+	public void setSize(int size) {
+		this.pro_size = size;
+	}
+	public int getColor() {
+		return color;
+	}
+	public void setColor(int color) {
+		this.color = color;
+	}
+	public int getCount() {
+		return count;
+	}
+	public void setCount(int count) {
+		this.count = count;
+	}
+
+	
+	
+}
